pyproject.toml
[project]
name = "socio-cognitive-algorithm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "contourpy==1.3.0",
    "cycler==0.12.1",
    "fonttools==4.54.1",
    "jmetalpy==1.7.0",
    "kiwisolver==1.4.7",
    "matplotlib==3.9.2",
    "mockito==1.5.3",
    "numpy==2.1.3",
    "packaging==24.2",
    "pandas==2.2.3",
    "patsy==0.5.6",
    "pillow==11.0.0",
    "plotly==5.24.1",
    "pyhamcrest==2.1.0",
    "pyparsing==3.2.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.2",
    "scipy==1.14.1",
    "six==1.16.0",
    "statsmodels==0.14.4",
    "tenacity==9.0.0",
    "tqdm==4.67.0",
    "tzdata==2024.2",
]
